Added support for two markup engines
There are now two parsers available.  I had to rename the source files
to avoid confusion between the parsers and the Python modules they use.

Now the user can specify the `markup` parameter as either `mikasa` or
`markdown`.  These are then the names of the respective Python
modules that will be used.

The advantage of using the `markdown` module is that it has nice
integration with Pygents which means syntax highlighting is quite nice.

There is a lingering issue which is that if you use the `markdown`
module, any `----` sequences get converted to `<hr/>` elements which
seems to confuse the code that (somewhere) cuts the HTML content
into slides.  This should also be a problem for `misaka` as well
but doesn't appear to be (which implies there is a bug in `misaka`).

I'll look into that next.