Fixed splitting issue
When switching between misaka and markdown processors, I noticed that the
slides didn't get split properly.  The issue was that the logical in
slides.py assumed HTML.  But markdown was spitting out XHTML.  The
difference was in the handling of the horizontal rule tag (and not
because of a bug in misaka as I earlier suspected).  So now the
splitting is robust whether the element is written as '<hr>', '<hr/>'
or '<hr />'.